requirements.txt
asgiref==3.5.2
coreapi==2.3.3
Django==4.0.6
django-ckeditor==6.4.2
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
drf-yasg==1.21.3
gunicorn==20.1.0
heroku==0.1.4
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
requests==2.28.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.11
whitenoise==6.2.0
cloudinary==1.29.0
